Abstract out calendar component for reuse (#55)
* Abstract out the calendar element

* Fix calendar outline missing
