+++ frontend/src/styles/styles.scss
@@ -544,7 +544,7 @@ body {
     border-bottom: 1px mat.get-theme-color($theme, outline-variant) solid;
   }
 
-  .operating-hours-calendar-outline {
+  .calendar-outline {
     border: 2px mat.get-theme-color($theme, outline-variant) solid;
   }
